pumukit: fix delete videos for restores.

+++ store/pumukit/classes/sessions.php
@@ -80,12 +80,25 @@ class sessions {
         global $DB;
 
         $pumukitconf = $this->load_config($config->id);
-        $pumukitclient = new pumukit_handler($pumukitconf);
+        $code = $DB->get_field('hybridteachstore_pumukit', 'code', ['sessionid' => $htsession]);
+        
+        if (isset($pumukitconf) && isset($code)) {
+            // Only delete video from pumukit if there are not same video with another session, for example, restored from backup.
+            $sql = "SELECT * 
+                FROM {hybridteachstore_pumukit} pk
+                WHERE pumukit LIKE :code";
+            $othersession = $DB->get_records_sql($sql, ['code' => $code]);
 
-        // Get cohorts.
-        $rows = $DB->get_records('hybridteachstore_pumukit', ['sessionid' => $htsession]);
-        foreach ($rows as $row) {
-            $pumukitclient->deletefile($row->id);
+            if (count($othersession) > 1) {
+                 // No delete if there are other restores.
+            } else {
+                // Delete video in pumukit
+                $rows = $DB->get_records('hybridteachstore_pumukit', ['sessionid' => $htsession]);
+                $pumukitclient = new pumukit_handler($pumukitconf);
+                foreach ($rows as $row) {
+                    $pumukitclient->deletefile($row->id);
+                }
+            }
         }
 
         $DB->delete_records('hybridteachstore_pumukit', ['sessionid' => $htsession]);
